error handling

+++ frontend/src/app/components/users/users.component.ts
@@ -1,6 +1,9 @@
+import { ConfirmationDialog } from './../confirmation-dialog/confirmation.dialog';
 import { UserService } from './../../services/user.service';
 import { Component, OnInit } from '@angular/core';
 import { User } from 'src/app/models/user.model';
+import { MatDialog } from '@angular/material/dialog';
+import { NotificationService } from 'src/app/services/notification.service';
 
 @Component({
   selector: 'app-users',
@@ -10,7 +13,7 @@ import { User } from 'src/app/models/user.model';
 export class UsersComponent implements OnInit {
   users: User[] = []
 
-  constructor(private userService: UserService) {}
+  constructor(private userService: UserService, public dialog: MatDialog, private notificationService: NotificationService) {}
 
   ngOnInit(): void {
     this.userService.getAll().subscribe((users) => {
@@ -19,9 +22,17 @@ export class UsersComponent implements OnInit {
   }
 
   delete(userId: number) {
-    this.userService.delete(userId).subscribe((users) => {
-      // TODO - display a message to the user
-      this.users = users;
+    const dialogRef = this.dialog.open(ConfirmationDialog, {
+      data: 'Confirmer la suppression de l\'utilisateur ?'
+    })
+
+    dialogRef.afterClosed().subscribe(result => {
+      if(result) {
+        this.userService.delete(userId).subscribe((users) => {
+          this.notificationService.showSuccess('Utilisateur supprimé');
+          this.users = users;
+        });
+      }
     });
   }
